+++ examples/statz.zig
@@ -15,7 +15,7 @@ pub const logging_override: debug.Logging.Override = spec.logging.override.silen
 pub fn main(args: [][*:0]u8) !void {
     var rcd_buf: [4096]build.Record = undefined;
     for (args[1..]) |arg| {
-        const fd: u64 = try file.open(.{}, meta.manyToSlice(arg));
+        const fd: usize = try file.open(.{}, meta.manyToSlice(arg));
         for (rcd_buf[0..try file.read(.{ .child = build.Record }, fd, &rcd_buf)]) |rcd| {
             var array: mem.StaticString(4096) = undefined;
             array.undefineAll();
